+++ docs/changelog.rst
@@ -6,7 +6,7 @@ All notable changes to pyvergeos will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/>`_.
 
-[1.3.0] - 2026-05-21
+[1.2.3] - 2026-05-21
 --------------------
 
 Added